refactor(experiments): consolidate shared physics into experiments/physics_config.py
- Create experiments/physics_config.py with pure-Python PhysicsConfig dataclass
  and finite_difference_gradient (central differences, no torch dependency)
- Refactor backprop_advanced/physics.py into torch-only extensions that
  re-export PhysicsConfig and finite_difference_gradient from the shared module
- Remove duplicate PhysicsConfig definitions from bouncing_ball_benchmark.py
  and bouncing_ball_pipeline.py; import from shared module instead
- Replace finite_difference_gradient re-implementation in verify_gradient.py
  with import from shared module; update call site to use PhysicsConfig
- Replace compute_analytical_gradient in plot_determinism_results.py with
  a direct call to the shared finite_difference_gradient
