+++ experiments/backprop_advanced/physics.py
@@ -1,25 +1,31 @@
 """
-Differentiable bouncing ball physics (scalar and batched).
+Differentiable bouncing ball physics — PyTorch batched utilities.
 
-Standalone module — no Retriever dependency.
-Works with both scalar θ and batched θ [B].
+Extends experiments/physics.py with torch-based batch simulation
+and autograd-compatible contact resolution.
+No Retriever dependency.
 """
 
 from __future__ import annotations
 
-from dataclasses import dataclass
+import os
+import sys
 
 import torch
 
+# Shared pure-Python physics (PhysicsConfig, finite_difference_gradient)
+_experiments_dir = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+if _experiments_dir not in sys.path:
+    sys.path.insert(0, _experiments_dir)
+from physics_config import PhysicsConfig, finite_difference_gradient  # re-exported for callers
 
-@dataclass
-class PhysicsConfig:
-    g: float = 9.81       # gravity
-    e: float = 0.8        # restitution coefficient
-    dt: float = 0.01      # time step
-    T: int = 100          # simulation horizon
-    x_target: float = 0.5
-    x_init: float = 1.0
+__all__ = [
+    "PhysicsConfig",
+    "finite_difference_gradient",
+    "step_batch",
+    "simulate_batch",
+    "batch_loss",
+]
 
 
 def step_batch(
@@ -79,34 +85,3 @@ def batch_loss(
     """
     x_T = simulate_batch(theta_batch, cfg)
     return (x_T - cfg.x_target) ** 2
-
-
-def finite_difference_gradient(
-    theta_val: float,
-    cfg: PhysicsConfig,
-    eps: float = 1e-6,
-) -> tuple[float, float]:
-    """
-    Scalar finite-difference gradient dL/dθ at a single θ value.
-
-    Uses central differences: (L(θ+ε/2) - L(θ-ε/2)) / ε
-
-    Returns:
-        (gradient, loss_at_theta)
-    """
-    def simulate_loss(th: float) -> float:
-        x, v = cfg.x_init, th
-        for _ in range(cfg.T):
-            v_pred = v - cfg.g * cfg.dt
-            x_pred = x + v_pred * cfg.dt
-            if x_pred < 0.0:
-                x, v = 0.0, -cfg.e * v_pred
-            else:
-                x, v = x_pred, v_pred
-        return (x - cfg.x_target) ** 2
-
-    loss = simulate_loss(theta_val)
-    loss_plus = simulate_loss(theta_val + eps / 2)
-    loss_minus = simulate_loss(theta_val - eps / 2)
-    grad = (loss_plus - loss_minus) / eps
-    return grad, loss
